LayoutTests/fast/events/drag-select-when-zoomed-expected.txt

This test ensures that drag-selecting works correctly in scaled pages.

Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.
PASS firstRange.startOffset is 79
PASS firstRange.endOffset is 103
PASS successfullyParsed is true

TEST COMPLETE
